Make XMDS-app "support" $HOME paths containing spaces.
Virtualenv does not support installation in directories that contain spaces.  So if $HOME contains a space, that rules out ~/Library/XMDS as an installation directory.  The best alternative I've come up with is to install in a temporary directory that may get blown away on every restart.  Currently there's no warning to the user that this might happen, but at least now it will work for such users.